requirements.txt
streamlit
ibm-watson-machine-learning
openai
pinecone-client
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
